chore: moves datepicker stories to scenarios
