+++ src/datepicker/__tests__/stateful-datepicker.scenario.js
@@ -1,0 +1,15 @@
+/*
+Copyright (c) 2018 Uber Technologies, Inc.
+
+This source code is licensed under the MIT license found in the
+LICENSE file in the root directory of this source tree.
+*/
+// @flow
+
+import React from 'react';
+
+import {Unstable_StatefulDatepicker as StatefulDatepicker} from '../index.js';
+
+export const name = 'Stateful datepicker';
+
+export const component = () => <StatefulDatepicker />;
