--- src/datepicker/examples-list.js
@@ -1,19 +1,0 @@
-/*
-Copyright (c) 2018 Uber Technologies, Inc.
-
-This source code is licensed under the MIT license found in the
-LICENSE file in the root directory of this source tree.
-*/
-/* eslint-env node */
-/* eslint-disable flowtype/require-valid-file-annotation */
-
-module.exports = {
-  SIMPLE_EXAMPLE: 'Stateless calendar',
-  STATEFUL_EXAMPLE: 'Stateful calendar',
-  STATEFUL_IN_POPOVER: 'Stateful datepicker',
-  STATEFUL_RANGE_IN_POPOVER: 'Stateful range datepicker',
-  STATEFUL_RANGE_QUICK_SELECT: 'Stateful range calendar with quick select',
-  STATEFUL_DATEPICKER_QUICK_SELECT:
-    'Stateful range datepicker with quick select',
-  WITH_OVERRIDES: 'Datepicker with overrides',
-};
